refactor to get tests to pass
